+++ tests/unit/test_main.py
@@ -22,6 +22,7 @@ def test__parse_args__from():
         tagged_metadata=False,
         tag_branch=None,
         full_commit=False,
+        strict=False,
     )
     assert parse_args(["from", "git"]).vcs == "git"
     assert parse_args(["from", "git", "--tag-branch", "foo"]).tag_branch == "foo"
@@ -47,6 +48,7 @@ def test__parse_args__from():
     assert parse_args(["from", "any", "--full-commit"]).full_commit is True
     assert parse_args(["from", "any", "--debug"]).debug is True
     assert parse_args(["from", "any", "--tagged-metadata"]).tagged_metadata is True
+    assert parse_args(["from", "any", "--strict"]).strict is True
     assert parse_args(["from", "subversion", "--tag-dir", "foo"]).tag_dir == "foo"
 
     with pytest.raises(SystemExit):